requirements.txt
blist==1.3.6
cassandra-driver==2.1.0
futures==2.1.6
six==1.7.3
wsgiref==0.1.2
